Return fn to cancel in-progress jsonp request
When destroying a page component, for example a video player, it’s
important to be able to cancel in-flight jsonp requests so the callback
doesn’t get called and all references to the player will be nulled so
the player can get garbage collected quickly.

If the callback were to hang around for another few seconds, this would
delay GC and call a function that assumes the player is in a good state
when it’s not.